backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
python-dotenv
pinecone-client
requests
